erc20: make connector-generic tests
The tests that we had are specific to the ETH connector for the purpose of comparing to OpenZeppelin.  But I thought we should have connector-generic tests because they will (A) show that the contract works on Algorand once it adds box storage, and (B) I think it will be easier for tutorials based on the ERC20 with “normal” generic tests instead of tests that rely on the ethers library.
